# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharmacy Management System API",
        "description": "Complete API documentation for Pharmacy Management System including admin, student, teacher, and payment modules",
        "contact": {
            "email": "support@pharmacy.edu"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.pharmacy.edu",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/marks/student-marks-info-v1": {
            "post": {
                "tags": [
                    "Student Marks"
                ],
                "summary": "Get student marks information",
                "description": "Retrieve student marks information using fn_admin_getstudentmarksinfo_v2 stored procedure. Use student_id=0 for all students or specific student_id for individual student.",
                "operationId": "cd04f8906d56714127f8369202292a93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "admin_user_id",
                                    "student_id",
                                    "inst_code",
                                    "dept_code",
                                    "subject_code",
                                    "exam_year",
                                    "semester"
                                ],
                                "properties": {
                                    "admin_user_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "student_id": {
                                        "description": "0 for all students, non-zero for specific student",
                                        "type": "integer",
                                        "example": 11452
                                    },
                                    "inst_code": {
                                        "type": "string",
                                        "example": "JCG"
                                    },
                                    "dept_code": {
                                        "type": "string",
                                        "example": "PHARM"
                                    },
                                    "subject_code": {
                                        "type": "string",
                                        "example": "PHCE"
                                    },
                                    "exam_year": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "semester": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "string",
                                            "example": "1.0"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "marksId": {
                                                    "type": "integer",
                                                    "example": 100480
                                                },
                                                "semester": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "studentId": {
                                                    "type": "integer",
                                                    "example": 11452
                                                },
                                                "studentNo": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "department": {
                                                    "type": "string",
                                                    "example": "PHARM"
                                                },
                                                "examStatus": {
                                                    "type": "string",
                                                    "example": "INC"
                                                },
                                                "isEditable": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "marksStatus": {
                                                    "type": "string",
                                                    "example": "DRAFT"
                                                },
                                                "studentName": {
                                                    "type": "string",
                                                    "example": "MOSRAF HOSSAIN"
                                                },
                                                "studentRoll": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "subjectCode": {
                                                    "type": "string",
                                                    "example": "PHCE"
                                                },
                                                "studentRegNo": {
                                                    "type": "string",
                                                    "example": "PHARM232408597"
                                                },
                                                "externalMarks": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": true
                                                },
                                                "internalMarks": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "nullable": true
                                                },
                                                "studentEnrlType": {
                                                    "type": "string",
                                                    "example": "READMISSION"
                                                },
                                                "subjectCategory": {
                                                    "type": "string",
                                                    "example": "THEORY"
                                                },
                                                "externalMaxMarks": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": 80
                                                },
                                                "internalMaxMarks": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": 20
                                                },
                                                "subjectCategoryId": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized - Token missing or invalid"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Enter your authentication token",
                "name": "token",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}